refactor: package 구조 변경
common: 패키지 전체에서 공통으로 쓰일 패키지의 모음
 constants: 상수 모음
 error: 에러처리 모음
  exception: 예외 모음
 utils: 유틸성 클래스 모음
api : REST API 요청은 이쪽 패키지에서 담당
web : html 요청은 이쪽 패키지에서 담당
