+++ src/main/java/com/codessquad/qna/common/error/exception/QuestionNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codessquad.qna.error.exception;
+package com.codessquad.qna.common.error.exception;
 
 public class QuestionNotFoundException extends EntityNotFoundException {
