+++ src/main/java/com/codessquad/qna/web/question/QuestionRepository.java
@@ -1,4 +1,4 @@
-package com.codessquad.qna.question;
+package com.codessquad.qna.web.question;
 
 import org.springframework.data.repository.CrudRepository;
